added making dynamic predictions -- not working
